requirement.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-generator==1.10
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorclass==2.2.2
docopt==0.6.2
exceptiongroup==1.2.2
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
html2markdown==0.1.7
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.19
langchain-google-genai==2.0.4
langgraph==0.2.52
langgraph-checkpoint==2.0.5
langgraph-sdk==0.1.36
langsmith==0.1.143
msgpack==1.1.0
multidict==6.1.0
mysql-connector-python==9.1.0
orjson==3.10.11
outcome==1.3.0
packaging==24.2
pip-upgrader==1.4.15
pip3-autoremove==1.2.2
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
py-cord==2.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
schedule==1.2.2
selenium==4.26.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tenacity==9.0.0
terminaltables==3.1.10
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.6
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.17.2
